backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
click==8.2.1
colorama==0.4.6
databases==0.9.0
fastapi==0.116.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
